Refactor to support new project-support naming scheme.
