+++ src/index.ts
@@ -1,2 +1,5 @@
-import {folderTest} from "./folderTest";
+import {folderTest} from "./folderTest/folderTest";
+import Log from "./logger/Log";
+
 export {folderTest};
+export {Log};
