+++ test/folderTest/readTestsFromDisk.spec.d.ts
@@ -1,0 +1,1 @@
+export {};
