+++ test/folderTest/validateTests.spec.d.ts
@@ -1,0 +1,1 @@
+export {};
